+++ packages/headless/src/controllers/smart-snippet/headless-smart-snippet.ts
@@ -1,4 +1,5 @@
 import {buildController, Controller} from '../controller/headless-controller';
+import {Result} from '../../api/search/search/result';
 import {search, questionAnswering} from '../../app/reducers';
 import {loadReducerError} from '../../utils/errors';
 import {QuestionAnswerDocumentIdentifier} from '../../api/search/search/question-answering';
@@ -7,6 +8,7 @@ import {
   logDislikeSmartSnippet,
   logExpandSmartSnippet,
   logLikeSmartSnippet,
+  logOpenSmartSnippetSource,
   logOpenSmartSnippetFeedbackModal,
   logCloseSmartSnippetFeedbackModal,
   logSmartSnippetFeedback,
@@ -23,8 +25,27 @@ import {
   openFeedbackModal,
 } from '../../features/question-answering/question-answering-actions';
 import {QuestionAnsweringSection} from '../../state/state-sections';
+import {getResultProperty} from '../../features/result-templates/result-templates-helpers';
+import {pushRecentResult} from '../../features/recent-results/recent-results-actions';
+import {buildInteractiveResultCore} from '../core/interactive-result/headless-core-interactive-result';
 
 export type {QuestionAnswerDocumentIdentifier} from '../../api/search/search/question-answering';
+
+export interface SmartSnippetOptions {
+  /**
+   * The amount of time in milliseconds to wait before selecting the source after calling `source.beginDelayedSelect`.
+   *
+   * @defaultValue `1000`
+   */
+  selectionDelay?: number;
+}
+
+export interface SmartSnippetProps {
+  /**
+   * The options for the `Tab` controller.
+   */
+  options?: SmartSnippetOptions;
+}
 
 /**
  * The `SmartSnippet` controller allows to manage the excerpt of a document that would be most likely to answer a particular query .
@@ -68,6 +89,28 @@ export interface SmartSnippet extends Controller {
    * @param details - A personalized message from the end user about the relevance of the answer.
    */
   sendDetailedFeedback(details: string): void;
+  /**
+   * Selects the source, logging a UA event to the Coveo Platform if the source wasn't already selected before.
+   *
+   * In a DOM context, we recommend calling this method on all of the following events:
+   * * `contextmenu`
+   * * `click`
+   * * `mouseup`
+   * * `mousedown`
+   */
+  selectSource(): void;
+  /**
+   * Prepares to select the source after a certain delay, sending analytics if it was never selected before.
+   *
+   * In a DOM context, we recommend calling this method on the `touchstart` event.
+   */
+  beginDelayedSelectSource(): void;
+  /**
+   * Cancels the pending selection caused by `beginDelayedSelect`.
+   *
+   * In a DOM context, we recommend calling this method on the `touchend` event.
+   */
+  cancelPendingSelectSource(): void;
 }
 
 /**
@@ -106,21 +149,57 @@ export interface SmartSnippetState {
    * Determines if the feedback modal with the purpose of explaining why the end user disliked the snippet is currently opened.
    */
   feedbackModalOpen: boolean;
+  /**
+   * Provides the source of the smart snippet.
+   */
+  source?: Result;
 }
 
 /**
  * Creates a `SmartSnippet` controller instance.
  *
  * @param engine - The headless engine.
+ * @param props - The configurable `SmartSnippet` properties.
  * @returns A `SmartSnippet` controller instance.
  * */
-export function buildSmartSnippet(engine: SearchEngine): SmartSnippet {
+export function buildSmartSnippet(
+  engine: SearchEngine,
+  props?: SmartSnippetProps
+): SmartSnippet {
   if (!loadSmartSnippetReducers(engine)) {
     throw loadReducerError;
   }
 
   const controller = buildController(engine);
   const getState = () => engine.state;
+
+  const getResult = () => {
+    const {contentIdKey, contentIdValue} =
+      getState().search.response.questionAnswer.documentId;
+    return engine.state.search.results.find(
+      (result) => getResultProperty(result, contentIdKey) === contentIdValue
+    );
+  };
+
+  let lastSearchResponseId: string | null = null;
+  const interactiveResult = buildInteractiveResultCore(
+    engine,
+    {options: {selectionDelay: props?.options?.selectionDelay}},
+    () => {
+      const result = getResult();
+      if (!result) {
+        lastSearchResponseId = null;
+        return;
+      }
+      const {searchResponseId} = getState().search;
+      if (lastSearchResponseId === searchResponseId) {
+        return;
+      }
+      lastSearchResponseId = searchResponseId;
+      engine.dispatch(logOpenSmartSnippetSource(result));
+      engine.dispatch(pushRecentResult(result));
+    }
+  );
 
   return {
     ...controller,
@@ -136,9 +215,9 @@ export function buildSmartSnippet(engine: SearchEngine): SmartSnippet {
         liked: state.questionAnswering.liked,
         disliked: state.questionAnswering.disliked,
         feedbackModalOpen: state.questionAnswering.feedbackModalOpen,
+        source: getResult(),
       };
     },
-
     expand() {
       engine.dispatch(logExpandSmartSnippet());
       engine.dispatch(expandSmartSnippet());
@@ -170,6 +249,15 @@ export function buildSmartSnippet(engine: SearchEngine): SmartSnippet {
     sendDetailedFeedback(details) {
       engine.dispatch(logSmartSnippetDetailedFeedback(details));
       engine.dispatch(closeFeedbackModal());
+    },
+    selectSource() {
+      interactiveResult.select();
+    },
+    beginDelayedSelectSource() {
+      interactiveResult.beginDelayedSelect();
+    },
+    cancelPendingSelectSource() {
+      interactiveResult.cancelPendingSelect();
     },
   };
 }